refactor(inquiry): preps stateful view for authentication integration
